Refactor db connection for better code understandability
