+++ Coding phase/Backend/connection.js
@@ -1,0 +1,32 @@
+const mysql = require('mysql');
+
+let db;
+
+function handleDisconnect() {
+    db = mysql.createConnection({
+        host: 'localhost',
+        user: 'root',
+        password: 'Str0ngP@ssw0rd!',
+        database: 'InterHostelCart'
+    });
+
+    db.connect(err => {
+        if (err) {
+            console.error('Error connecting to database:', err);
+            setTimeout(handleDisconnect, 2000); // Reconnect after 2 seconds
+        }
+    });
+
+    db.on('error', err => {
+        console.error('Database error:', err);
+        if (err.code === 'PROTOCOL_CONNECTION_LOST') {
+            handleDisconnect(); // Reconnect if connection is lost
+        } else {
+            throw err;
+        }
+    });
+}
+
+handleDisconnect();
+
+module.exports = db;
